feat : move utils
