+++ backend/src/main/java/toonpick/app/controller/MemberFavoriteWebtoonController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import toonpick.app.common.utils.AuthenticationUtil;
+import toonpick.app.utils.AuthenticationUtil;
 import toonpick.app.service.MemberFavoriteWebtoonService;
 import toonpick.app.dto.WebtoonDTO;
