+++ backend/src/main/java/toonpick/app/domain/WatchingToons/WatchingToons.java
@@ -14,7 +14,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import toonpick.app.common.utils.BaseTimeEntity;
+import toonpick.app.domain.BaseTimeEntity;
 import toonpick.app.domain.member.Member;
 import toonpick.app.domain.webtoon.Webtoon;
