+++ backend/src/main/java/toonpick/app/utils/ErrorResponseSender.java
@@ -1,4 +1,4 @@
-package toonpick.app.common.utils;
+package toonpick.app.utils;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
